+++ build.gradle
@@ -26,7 +26,7 @@ buildscript {
                                 'org.mockito:mockito-core:2.8.47',
                                 'com.nhaarman.mockitokotlin2:mockito-kotlin:2.2.0',
                                 "org.jetbrains.kotlinx:kotlinx-coroutines-test:$kotlinx_coroutines_version",
-                                "io.mockk:mockk:1.9.3"
+                                "io.mockk:mockk:1.9.2"
                         ],
                         android: [
                                 'org.mockito:mockito-android:2.13.0',
